Convert padding from i32 to i64 to fix integer overflow
Fixes MNIST example crash caused by overflow when handling negative
indices in tensor slicing operations. Padding calculations now use i64
which provides sufficient range for indices up to -2147483648.

+++ zyx/src/compiled_graph.rs
@@ -59,7 +59,7 @@ pub enum CompiledNode {
 pub struct CompactedGraph {
     pub nodes: Vec<Node>,
     pub shapes: BTreeMap<TensorId, Box<[Dim]>>,
-    pub paddings: BTreeMap<TensorId, Box<[(i32, i32)]>>,
+    pub paddings: BTreeMap<TensorId, Box<[(i64, i64)]>>,
     pub axes: BTreeMap<TensorId, Box<[UAxis]>>,
 }
