+++ zyx/src/graph.rs
@@ -40,7 +40,7 @@ pub struct Graph {
     pub gradient_tape_ref_count: u32,
     pub gradient_tape: Option<Set<TensorId>>,
     pub shapes: Map<TensorId, Box<[Dim]>>,
-    paddings: Map<TensorId, Box<[(i32, i32)]>>,
+    paddings: Map<TensorId, Box<[(i64, i64)]>>,
     axes: Map<TensorId, Box<[UAxis]>>,
 }
 
@@ -130,7 +130,7 @@ impl Graph {
         id
     }
 
-    pub(super) fn push_padding(&mut self, id: TensorId, padding: Vec<(i32, i32)>) {
+    pub(super) fn push_padding(&mut self, id: TensorId, padding: Vec<(i64, i64)>) {
         self.paddings.insert(id, padding.into_boxed_slice());
     }
 
@@ -189,7 +189,7 @@ impl Graph {
         panic!("DType of {tensor_id:?} could not be found. This is internal bug.")
     }
 
-    pub(super) fn padding(&self, tensor_id: TensorId) -> &[(i32, i32)] {
+    pub(super) fn padding(&self, tensor_id: TensorId) -> &[(i64, i64)] {
         &self.paddings[&tensor_id]
     }
